Fix geolocation on the web
I noticed that clicking "get geolocation" in the web app settings results in "not implemented" error, and on the setup page it's hidden altogether (presumably due to the same error in the past) and thought it might be fun to dig in.

Looks like this is because requestPermissions is not supported on the web in Capacitor, so as long as we skip it on this platform, everything else works fine.

In the process I noticed a lot of duplicated code for geolocation between setup and settings pages (which makes sense, the logic is pretty much the same) so I extracted it out into the utils.